fix: sidekick no longer opening in dev
